pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "promptfuzz"
version = "0.1.3"
description = "Adversarial prompt fuzzing framework for LLM applications"
readme = "README.md"
license = { text = "Apache-2.0" }
requires-python = ">=3.11"
authors = [{ name = "Your Name", email = "you@example.com" }]
keywords = ["llm", "security", "fuzzing", "adversarial", "ai-safety", "red-team"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

dependencies = [
    "click>=8.1.8",
    "rich>=14.3.3",
    "pyyaml>=6.0.3",
    "pydantic>=2.12.5",
    "httpx>=0.28.1",
    "litellm>=1.83.3",
    "tenacity>=9.1.4",
    "python-dotenv>=1.0.1",
    "fastapi>=0.111.0",
    "uvicorn>=0.30.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "mypy>=1.20.0",
    "ruff>=0.15.9",
    "pre-commit>=4.5.1",
    "types-PyYAML>=6.0.12",
]
all = ["promptfuzz[dev]"]

[project.scripts]
promptfuzz = "promptfuzz.cli:main"

[project.urls]
Homepage = "https://github.com/yourname/promptfuzz"
Documentation = "https://promptfuzz.dev"
Repository = "https://github.com/yourname/promptfuzz"
Issues = "https://github.com/yourname/promptfuzz/issues"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B", "SIM", "TCH"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=promptfuzz --cov-report=term-missing"

[tool.mypy]
disable_error_code = ["type-abstract", "no-untyped-def", "no-untyped-call"]

python_version = "3.11"
strict = true
